[major] Dynamically generate and control VM inventory
This change results in a locally generated inventory of virtual machines
that have been built from the virt-host-setup.yml playbook. You can then
use this dynamic inventory file for the steps beyond the virtual host setup.

Updated the documentation as well showing some of the new variables that
I added, which allows for the usage of a jump host via the SSH Proxy command
and local SSH key, along with variables to define the virthost location.