Add new role to gather kube niceties
Add a new role that we can use to add some console niceties to
our deployments. The first one is just adding kubectl bash completion
to the main user on the master node.

In the future we could add something like c-bata/kube-prompt or other
things that make the working environment more enjoyable to
be in.

Closes #62