Load virthost privkey onto source machine
This change adds the functionality to pull the remotely created SSH
private key from the virtual host onto the source host. It does this
via the fetch module which retrieves the generated root SSH private
key so that authentication can be done when running the
kube-install.yml playbook.

We place the key into the source machines
$HOME/.ssh/{{ inventory_host }}/ directory so that we don't conflict
with other local files.

Additionally, in order to make things cleaner and more obvious, I have
renamed the kubehost group to virthost and performed some other minor
cleanup. Because I'm a bit of a badass like that.

Closes #61