Enhance documentation and usage experience
In this set of changes, we fix up the README to show a lot more
of the proper usage and setup of kucean. A new topology graphic
has been added that helps to make it clearer the expected
topology being utilized.

Additionally, default to using the NAT'd KVM bridge and not
a direct virthost bridge, since that would require additional
setup by the developer that might not be expected in a
development environment.

The ansible configuration is now setup to ignore the deprecated
usage of `include` until we migrate everything over to the new
`import_...` usage.

Cleaned up `vm-spinup` role so that you can likely just define
the `ssh_proxy_enabled` to true, and likely get what you expect
in the `vms.local.generated` file.

Closes #87