Copies artifacts from the builder to the k8s nodes
This change moves the artifacts built from a k8s build over to the k8s
nodes (master and nodes groups).

In order to allow the synchronize module (rsync) to work properly with
the SSH keys, we need to execute this as a non-root user to copy the
files over. When we do this, the files are placed into a temporary
k8s_artifacts directory.

Once the sync is completed, then a separate set of tasks is run to
copy the files on the remote node into their final destination, set
the proper owner, and then delete the temporary k8s_artifacts
directory from the /tmp/ directory.

Closes #119