[enhance] Only sync artifacts from list if they exist
Instead of blindly attempting to synchronize artifacts from the
archive_list, check if they are available, and build a list
of the artifacts to sync.

Also provides a fail module output for any of artifacts we expect
to exist, that don't, so it's obvious that things aren't as we
wish them to be. It's likely that if we don't sync all the files
over we expect, that our deploy will fail.

Because this change ends up making the synchronization process quite
a bit noisier when skipping over files, we've migrated to using the
skippy stdout callback module to no longer show skipped commands.

Closes #140