Refactor playbook location to match openshift-ansible
Refactor the playbooks location and layout to better match how
openshift-ansible is structured. The purpose of this is so that moving
from the kube-ansible lab environment to an OpenShift-Ansible
staging/production environment is more intuitive.

Note that the original group_vars/ values have now been moved to the
ka-init/ role, and that you should override variables from the
inventory file, or via an overrides yml file (-e
"./inventory/override-vars.yml").

Closes #183