[monitoring] Add Prometheus-Operator role and playbooks
Adds cluster monitoring using Prometheus-Operator, which is deployed using
kube-prometheus. Clones the repository, installs, and sets up the the
monitoring system for you. Also provides an uninstall method.

Uses the openshift-ansible format of playbooks which provides a post-install
task, but also designed to be consumed from a wrapper playbook that can
dynamically include the private/ playbooks if necessary (not yet added).

Closes #167