Provide config knobs to specify ovn-kubernets repo and branch
kube-ansible project clones the ovn-kubernetes project
and build the project to generate the OVN deployment yaml
files. Currently it's hardcoded to clone the official
ovn-kubernets repo and build the master branch by default.
Currently there is no way user can test it's work-in-progress
patch for ovn-kubernetes, specifically if that patch changes
the code that is not part of the ovn-kubernetes image, which
user can build from any repo and branch, like updating the
deployment yaml files.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>